+++ cypress/locators/confirm/index.js
@@ -1,4 +1,3 @@
-import { CLOSE_ICON_BUTTON } from "../locators";
 import {
   DIALOG,
   CANCEL_BUTTON,
@@ -8,7 +7,6 @@ import {
 
 // component preview locators
 export const dialogPreview = () => cy.get(DIALOG);
-export const closeIconButton = () => cy.get(CLOSE_ICON_BUTTON);
 export const dialogSubtitle = () => cy.get(DIALOG_SUBTITLE);
 export const confirmButton = () => cy.get(CONFIRM_BUTTON);
 export const cancelButton = () => cy.get(CANCEL_BUTTON);