+++ cypress/locators/popover-container/locators.js
@@ -5,6 +5,5 @@ export const POPOVER_CONTAINER_CONTENT =
   '[data-element="popover-container-content"]';
 export const POPOVER_CONTAINER_TITLE =
   '[data-element="popover-container-title"]';
-export const POPOVER_SETTINGS_ICON = '[data-element="settings"]';
 export const POPOVER_CONTENT_CLOSE_ICON =
   '[data-element="popover-container-close-component"]';