+++ cypress/locators/table/index.js
@@ -4,10 +4,8 @@ import {
   TABLE,
   TABLE_HEADER,
   PAGINATION_BUTTON,
-  ACTION_TOOLBAR,
   TABLE_AJAX,
 } from "./locators";
-import { BUTTON_DATA_COMPONENT } from "../pages/locators";
 
 // component preview locators
 export const rows = () => cy.get(ROW).parent();
@@ -21,15 +19,6 @@ export const sortIcon = (index) => tableHeader(index).find("span");
 export const pagination = () => cy.get(PAGINATION_BUTTON);
 export const paginationButtonByIndex = (index) =>
   pagination().find("div:nth-child(2) > button").eq(index);
-export const actionToolbar = (index) =>
-  cy.get(ACTION_TOOLBAR).find("div:nth-child(2) > div").eq(index);
-export const actionToolbarButton = () =>
-  cy
-    .get(ACTION_TOOLBAR)
-    .find("div:nth-child(2)")
-    .find(BUTTON_DATA_COMPONENT)
-    .parent();
-export const checkboxInHeader = () => cy.get(CHECKBOX_CELL);
 export const tableBody = () => cy.get(TABLE).find("table > tbody");
 export const tableAjax = () => cy.get(TABLE_AJAX).find("table > tbody");